Adding code to prevent checkins on parent branches from being applied to git repo.
Often a view will be configured as follows:

element * /main/MY_BRANCH/LATEST
mkbranch MY_BRANCH
element * /main/LATEST
endbranch

Take the example where you have PetStore.java@@/main/MY_BRANCH/4 in
your view.  Let's say your fried, Joe, commits PetStore.java@@/main/5.
As far as your view is concerned, Joe's commit is superfluous as it
is filtered out by the first rule "element * /main/MY_BRANCH/LATEST"
in your configspec.  Even though Joe's checkin creats a new item
in the element's history, we want to prevent Joe's checkin from
entering our Git repo, since it will not be part of your view.

To do this, the following logic can be used.  If a checkin of a file
is found in clearcase's history during a rebase that is *NOT* on
the same branch as the version of the file that currently exists in
your view, AND the checkin occured *AFTER* the most recent modification
of the file that exists in your view, then said checkin should be ignored,
as it does not apply directly to your view.

Note that this logic still allows for the history of elements to be
imported from parent branches UP TO the point that the first node was
created on the branch which your view is configured to pull from.

To accomplish this in a somewhat effecient fashion, I gather the version
information for all elements that exist in the view (filtering on
the folders configured for gitcc inclusion, of course) into a dicitonary,
using the file name and path as the key.  If required, the version
dictionary is then referenced later when the current version of an
element in the view must be evaluated.

This is most likely faster then making a call to "cleartool desc" each
time the version information is needed, depending on how many checks
are needed.  For the purposes of an initial rebase that imports a
significant amount of history, this method of collecting all of
the current version information is most definitely more efficient.