Fixing issues with checkin times
It is possible that a directory element is checked in before
the file element that has been added to the directory.  Thus,
if we can't find any history fo the newly added element before
the directory has been added, then we check for the first checked-in
element that appears in the element's history after the
directory is checkedin.

Also, since clearcase reports timezone offsets in lshistory, we
have to adjust times accordingly before using their values in
comparisons.