Fixing stupid ass Clearcase timezone issue with lshistory -since.
Clearcase is such a piece of shit, that if don't explicitly assign a timezone, it completly ignores the time zone when retrieving history.

Let's say you are in centratl time (CST, -0600).  If you do not explicitly provide lshistory with a UTC offset when passing in an arugment with the -since flag, it will incorrectly return checkins that occured in the past, if they occured in a timezone that is "ahead" of yours.  For example, in my case, where I am in central time, if I provide -since 14:05:00 to lshistory, it will return anything after 14:05:00, including a checking at 15:05-0500 (EST -0500).   WTF!?!?!?

Anyway, this patch ensures that a UTC offset is always used when formating dates for use with lshistory.