Added recovery, and auto merging capability
Auto Recovery:
==============
The script can now attempt to recover if it fails at some point during
the middle of a synchronization attempt.  Although not fool proof, it
is fairly good about figuring out where it was, and starting from
that point.

Auto Merge
==============
When files have changed in both clearcase and git, the script will try to
merge in the changes automagically using git merge-files.  If this fails,
it will attempt to open up a merge tool (current Perforce's mergetool) to
allow the user to resolve the merge conflicts.  It will then continue on
it's merry way.

Other Fixes
==============
* Better output; stacktraces are dumped when possible.
* Numerous minor issues/typos fixed.
* Prompts added to alert user when bad things could happen, and to
give the user options.